fix import errors from refactor, code compiles
